extract file icons flow to its own file
