+++ client/web/src/repo/utils.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from 'vitest'
 
-import { FileExtension } from './constants'
-import { containsTest, getFileInfo, getInitialSearchTerm } from './utils'
+import { FileExtension, containsTest, getFileInfo } from './fileIcons'
+import { getInitialSearchTerm } from './utils'
 
 describe('containsTest', () => {
     const tests: {
