+++ client/web/src/repo/utils.tsx
@@ -1,6 +1,6 @@
 import { type GitCommitFields, RepositoryType } from '../graphql-operations'
 
-import { CodeHostType, FileExtension } from './constants'
+import { CodeHostType } from './constants'
 
 export const isPerforceChangelistMappingEnabled = (): boolean =>
     window.context.experimentalFeatures.perforceChangelistMapping === 'enabled'
@@ -44,50 +44,4 @@ export const stringToCodeHostType = (codeHostType: string): CodeHostType => {
             return CodeHostType.OTHER
         }
     }
-}
-
-export interface FileInfo {
-    extension: FileExtension
-    isTest: boolean
-}
-
-export const getFileInfo = (file: string, isDirectory: boolean): FileInfo => {
-    if (isDirectory) {
-        return {
-            extension: 'default' as FileExtension,
-            isTest: false,
-        }
-    }
-
-    const extension = file.split('.').at(-1)
-    const isValidExtension = Object.values(FileExtension).includes(extension as FileExtension)
-
-    if (extension && isValidExtension) {
-        return {
-            extension: extension as FileExtension,
-            isTest: containsTest(file),
-        }
-    }
-
-    return {
-        extension: 'default' as FileExtension,
-        isTest: false,
-    }
-}
-
-export const containsTest = (file: string): boolean => {
-    const f = file.split('.')
-    // To account for other test file path structures
-    // adjust this regular expression.
-    const isTest = /^(test|spec|tests)(\b|_)|(\b|_)(test|spec|tests)$/
-
-    for (const i of f) {
-        if (i === 'test') {
-            return true
-        }
-        if (isTest.test(i)) {
-            return true
-        }
-    }
-    return false
 }
